+++ fishbowl_common/__init__.py
@@ -1,14 +1,40 @@
 from fishbowl_common.ArgumentProvider import ArgumentProvider
 from fishbowl_common.PatchNotes import PatchNotes
 from fishbowl_common.SettingsRepository import SettingsRepository
-from fishbowl_common.UpdateChecker import ReleaseAsset, UpdateChecker, UpdateCheckResult
+from fishbowl_common.UpdateChecker import (
+    CHECK_ERROR_HTTP,
+    CHECK_ERROR_NETWORK,
+    CHECK_ERROR_RATE_LIMITED,
+    CHECK_ERROR_RESPONSE,
+    ReleaseAsset,
+    UpdateChecker,
+    UpdateCheckResult,
+)
 from fishbowl_common.UpdateCoordinator import UpdateCoordinator, UpdateDisplay
-from fishbowl_common.UpdateDownloader import UpdateDownloader
+from fishbowl_common.UpdateDownloader import (
+    DOWNLOAD_ERROR_DIGEST,
+    DOWNLOAD_ERROR_HTTP,
+    DOWNLOAD_ERROR_IO,
+    DOWNLOAD_ERROR_NETWORK,
+    DOWNLOAD_ERROR_NO_DIGEST,
+    DOWNLOAD_ERROR_SIZE,
+    UpdateDownloader,
+)
 from fishbowl_common.UpdateInstaller import UpdateInstaller
 from fishbowl_common.version_utils import compare_versions, parse_version
 
 __all__ = [
     "ArgumentProvider",
+    "CHECK_ERROR_HTTP",
+    "CHECK_ERROR_NETWORK",
+    "CHECK_ERROR_RATE_LIMITED",
+    "CHECK_ERROR_RESPONSE",
+    "DOWNLOAD_ERROR_DIGEST",
+    "DOWNLOAD_ERROR_HTTP",
+    "DOWNLOAD_ERROR_IO",
+    "DOWNLOAD_ERROR_NETWORK",
+    "DOWNLOAD_ERROR_NO_DIGEST",
+    "DOWNLOAD_ERROR_SIZE",
     "PatchNotes",
     "ReleaseAsset",
     "SettingsRepository",